Fix ios framework build, reorganize ios tutorials and introduce a new tutorial to show how to use apriltag on ios

+++ modules/io/src/video/vpDiskGrabber.cpp
@@ -378,7 +378,7 @@ vpDiskGrabber::setImageNumber(long number)
   Set the step between two images.
 */
 void
-vpDiskGrabber::setStep(int step)
+vpDiskGrabber::setStep(long step)
 {
   m_image_step = step;
 }
